+++ app/src/main/java/ad/auction/dashboard/model/calculator/calculations/CPC.java
@@ -37,7 +37,7 @@ public class CPC extends Metric {
             try {
                 for (int i = 1; i < totalCost.get().size(); i++) {
                     var tcst = totalCost.get().get(i);
-                    points.add(new Point2D(tcst.getX(), (double) tcst.getY() / clickCount.get().get(i).getY()));
+                    points.add(new Point2D(tcst.getX(), tcst.getY() / clickCount.get().get(i).getY()));
                 }
             } catch (Exception ignored) {}
